+++ src/math/Point2D.h
@@ -9,15 +9,15 @@ extern "C" {
 
 #include "parsers/types.h"
 
-Point2D* point_2d_new(ssize_t x, ssize_t y);
+Point2D* point_2d_new(int32_t x, int32_t y);
 
 void point_2d_free(Point2D* p);
 
-Point2DArray* point_2d_array_new(Point2D* points, size_t length);
+Point2DArray* point_2d_array_new(Point2D* points, uint32_t length);
 
-Point2DArrayGrid* point_2d_array_grid_new(size_t width, size_t height, size_t times);
+Point2DArrayGrid* point_2d_array_grid_new(uint32_t width, uint32_t height, uint32_t times);
 
-Point2DArray* point_2d_array_new_empty(size_t length);
+Point2DArray* point_2d_array_new_empty(uint32_t length);
 
 void point2d_array_print(const Point2DArray* array);
