+++ src/matrix/tensor.h
@@ -8,17 +8,17 @@
 extern "C" {
 #endif
 
-Tensor* tensor_new(size_t width, size_t height, size_t depth);
+Tensor* tensor_new(uint32_t width, uint32_t height, uint32_t depth);
 
-TensorSet* tensor_set_new(size_t count, Tensor** tensors);
+TensorSet* tensor_set_new(uint32_t count, Tensor** tensors);
 
 void tensor_set_free(TensorSet* set);
 
 bool tensor_equals(const Tensor* t1, const Tensor* t2);
 
-Vector2D* get_dir_kernel(ssize_t D, ssize_t size);
+Vector2D* get_dir_kernel(int32_t D, int32_t size);
 
-Vector2D* vector2d_clone(const Vector2D* src, size_t len);
+Vector2D* vector2d_clone(const Vector2D* src, uint32_t len);
 
 void free_Vector2D(Vector2D* vec);
 
@@ -26,22 +26,22 @@ void tensor_free(Tensor* tensor);
 
 Tensor* tensor_copy(const Tensor* original);
 
-void tensor_fill(Tensor* tensor, double value);
+void tensor_fill(Tensor* tensor, float value);
 
-int tensor_in_bounds(Tensor* tensor, size_t x, size_t y, size_t z);
+int tensor_in_bounds(Tensor* tensor, uint32_t x, uint32_t y, uint32_t z);
 
 Tensor* tensor_clone(const Tensor* src);
 
-size_t tensor_save(Tensor* tensor, const char* foldername);
+uint32_t tensor_save(Tensor* tensor, const char* foldername);
 
 Tensor* tensor_load(const char* foldername);
 
 typedef struct {
-    size_t len_data;
+    uint32_t len_data;
     Tensor** data;
 } Tensor4D;
 
-void tensor4D_free(Tensor** tensor, ssize_t T);
+void tensor4D_free(Tensor** tensor, int32_t T);
 
 
 #ifdef __cplusplus
